refact fleet_moe.
