+++ python/paddle/distributed/__init__.py
@@ -64,7 +64,6 @@ from paddle.fluid.dygraph.parallel import ParallelEnv  # noqa: F401
 
 from . import cloud_utils  # noqa: F401
 from . import utils  # noqa: F401
-
 from .sharding import *  # noqa: F401
 
 __all__ = [  # noqa
