+++ python/paddle/distributed/fleet/moe/gate/naive_gate.py
