+++ python/paddle/distributed/fleet/moe/moe_layer.py
@@ -21,12 +21,11 @@
 
 import collections
 import math
-
 import numpy as np
 import paddle
 import paddle.nn as nn
 import paddle.nn.functional as F
-from paddle.distributed.utils import global_scatter, global_gather
+from ...utils import global_scatter, global_gather
 from paddle.distributed import alltoall, all_gather
 
 from paddle.distributed.fleet.meta_parallel import get_rng_state_tracker
