+++ python/paddle/fluid/tests/unittests/test_number_count_op.py
@@ -23,7 +23,7 @@ from paddle.fluid.op import Operator
 import paddle.fluid as fluid
 from paddle.fluid import compiler, Program, program_guard
 from paddle.fluid.backward import append_backward
-from paddle.distributed.models.moe import utils
+from paddle.distributed.fleet.moe import utils
 from paddle.fluid.framework import _test_eager_guard
