+++ python/paddle/fluid/tests/unittests/test_prune_gate_by_capacity_op.py
@@ -15,7 +15,7 @@
 import unittest
 import paddle
 import numpy as np
-from paddle.distributed.models.moe import utils
+from paddle.distributed.fleet.moe import utils
 from paddle.fluid import core
 from paddle.fluid.framework import _test_eager_guard
