+++ python/setup.py.in
@@ -314,8 +314,6 @@ packages=['paddle',
           'paddle.distributed.auto_parallel.tuner',
           'paddle.distributed.auto_parallel.cost',
           'paddle.distributed.passes',
-          'paddle.distributed.models',
-          'paddle.distributed.models.moe',
           'paddle.framework',
           'paddle.jit',
           'paddle.jit.dy2static',
@@ -371,10 +369,7 @@ packages=['paddle',
           'paddle.incubate.nn.layer',
           'paddle.incubate.optimizer.functional',
           'paddle.incubate.autograd',
-          'paddle.incubate.distributed',
-          'paddle.incubate.distributed.models',
-          'paddle.incubate.distributed.models.moe',
-          'paddle.incubate.distributed.models.moe.gate',
+          'paddle.distributed.fleet.moe',
           'paddle.incubate.sparse',
           'paddle.incubate.sparse.nn',
           'paddle.incubate.sparse.nn.layer',
